requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pyjwt
passlib[bcrypt]
python-multipart